+++ PowerShell/Set-UserRightsAssignment/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Set-UserRightsAssignment release history
 
+## 1.0.1 - 2022-09-15
+
+* Added `Replace` option.
+* Set `ErrorActionPreference` on a single line.
+
 ## 1.0.0 - 2022-09-14
 
 ### First version